Size is not grandeur, and territory does not  make a nation

+++ 0x00-hello_world/6-size.c
@@ -9,8 +9,8 @@ int main(void)
 {
 	printf("Size of a char: %lu byte(s)\n", (unsigned long int) sizeof(char));
 	printf("Size of an int: %lu byte(s)\n", (unsigned long int) sizeof(int));
-	printf("Size of a long int: %lu byte(s)\n", (unsigned long int) sizeof(long int));
-	printf("Size of a long long int: %lu byte(s)\n", (unsigned long int) sizeof(long long int));
+	printf("Size of a long int: %lu byte(s)\n", (int) sizeof(long int));
+	printf("Size of a long long int: %lu byte(s)\n", (int) sizeof(long long int));
 	printf("Size of a float: %lu byte(s)\n", (unsigned long int) sizeof(float));
 	return (0);
 }
